Move SequenceRecord tests to separate module
